strkit: add MarkedSpec comment trimming
Motivated by a need to annotate intended trailing space in an input
string:

    > such input // NOTE we meant that space
    // which would have otherwise been invisible

But since we'd already allowed code-fenced value strings, stripping
within such is contraindicated; also it was always a bug to be skipping
comment lines in such.

So, surprise! Spec line scanning now needs to be a small state machine.